+++ scripts/generate-data.sh
@@ -50,6 +50,7 @@ pnpm exec tsx scripts/utils/src/update-tp2-highlight.ts \
     -i "$data_dir/weidu-baf-base.yml" -i "$data_dir/weidu-baf-iesdp.yml" -i "$data_dir/weidu-baf-ids.yml" \
     --completion "$dest_dir/completion.weidu-baf.json" \
     --hover "$dest_dir/hover.weidu-baf.json" \
+    --strrefs "$dest_dir/strrefs.weidu-baf.json" \
     --tooltip-lang weidu-baf-tooltip
 
 pnpm exec tsx scripts/utils/src/update-baf-highlight.ts \